services & users working

+++ roles/importer/files/importer/import_mgm.py
@@ -82,6 +82,7 @@ jwt = fwo_api.login(importer_user_name, importer_pwd, user_management_api_base_u
                     method, ssl_verification=ssl_mode, proxy_string=proxy_setting)
 
 # set import lock
+# todo: read url from config
 fwo_api_base_url = 'https://localhost:9443/api/v1/graphql'
 query_variables = {"mgmId": int(args.mgm_id)}
 current_import_id = fwo_api.lock_import(fwo_api_base_url, jwt, query_variables)
@@ -90,12 +91,11 @@ if current_import_id == -1:
                  str(args.mgm_id) + ", import already running?")
     sys.exit(1)
 
-logging.info("start import of management " + str(args.mgm_id) +
-             ", import_id=" + str(current_import_id))
+logging.info("start import of management " + str(args.mgm_id) + ", import_id=" + str(current_import_id))
 # from here on we have an import lock and need to unlock it before exiting
 
-mgm_details = fwo_api.get_mgm_details(fwo_api_base_url, jwt, query_variables)[
-    'data']['management'][0]  # get mgm_details (fw-type, port, ip, user credentials)
+# get mgm_details (fw-type, port, ip, user credentials):
+mgm_details = fwo_api.get_mgm_details(fwo_api_base_url, jwt, query_variables)['data']['management'][0] 
 
 full_config_json = {}
 config2import = {'rulebases': [], 'users': {}}
@@ -126,17 +126,14 @@ if mgm_details['deviceType']['name'] == 'Check Point' and mgm_details['deviceTyp
     os.system(get_config_cmd)
     with open(config_filename, "r") as json_data:
         full_config_json = json.load(json_data)
-    checkpointR8x.parse_network.parse_network_objects(
-        full_config_json, config2import, current_import_id)
-    checkpointR8x.parse_service.parse_service_objects(
-        full_config_json, config2import)
-    users = {}
+    checkpointR8x.parse_network.parse_network_objects(full_config_json, config2import, current_import_id)
+    checkpointR8x.parse_service.parse_service_objects(full_config_json, config2import, current_import_id)
+    users = []
     rulebase_range = range(len(rulebase_string.split(','))-1)
     for i in rulebase_range:
-        checkpointR8x.parse_user.collect_users_from_rulebase(
-            full_config_json['rulebases'][i], users)
+        checkpointR8x.parse_user.collect_users_from_rulebase(full_config_json['rulebases'][i], users)
         #checkpointR8x.parse_rule.parse_rulebase(full_config_json, config2import, rulebase)
-    config2import['users'].update(users)
+    config2import.update({'users': users})
     # config2import['rulebases'][rulebase].update)
 
 if mgm_details['deviceType']['name'] == 'Fortinet' and mgm_details['deviceType']['version'] == '5.x-6.x':
@@ -150,25 +147,20 @@ error_count += fwo_api.import_json_config(fwo_api_base_url, jwt, args.mgm_id, {
 
 change_count = fwo_api.count_changes_per_import(fwo_api_base_url, jwt, current_import_id)
 
-if change_count > 0:
-    # store full config
+if change_count > 0 or error_count>0: # store full config
     with open(config_filename, "r") as json_data:
         full_config_json = json.load(json_data)
 
     error_count += fwo_api.store_full_json_config(fwo_api_base_url, jwt, args.mgm_id, {
         "importId": current_import_id, "mgmId": args.mgm_id, "config": full_config_json})
 
-else:
-    # delete imports without changes:
+if change_count > 0 and error_count==0: # delete imports without changes (if no error occured)
     error_count += fwo_api.delete_json_config(fwo_api_base_url, jwt, {"importId": current_import_id} )
-    # read json file and write to FWO database via FWO API
-
-
 
 stop_time = int(time.time())
 stop_time_string = datetime.datetime.now().isoformat()
 
-error_count += fwo_api.unlock_import(fwo_api_base_url, jwt, int(args.mgm_id), stop_time_string, current_import_id, error_count, change_count)
+error_count += fwo_api.unlock_import(fwo_api_base_url, jwt, int(args.mgm_id), stop_time_string, current_import_id, error_count, change_count>0)
 
 if error_count:
     logging.warning("import ran with errors, duration: " +
